+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Upgraded codemirror from version 5 to version 6
 - Add code editor to allow editing the entire style
 - Add support for sprite object in setting modal
+- Allow root-relative urls in the stylefile
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes